Changed MVVM responsibilities and cleaned up the code

+++ src/Application/MVVM/View/CharacterSheet/CharacterViewModel.java
@@ -1,9 +1,10 @@
 package Application.MVVM.View.CharacterSheet;
 
 import Application.Client.ClientCharacterSheet;
-import Application.MVVM.Model.CharacterSheet.ICharacterSheetModel;
+
 import Application.MVVM.Model.character.Character;
 import javafx.beans.property.*;
+import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.scene.control.MultipleSelectionModel;
 import javafx.scene.paint.Color;
@@ -34,14 +35,16 @@ public class CharacterViewModel
   private final StringProperty saveStatusText;
   private final ObjectProperty<Paint> saveStatusColor;
 
+  private final BooleanProperty removeCharacterButtonDisabled;
+  private final BooleanProperty characterInfoVisible;
+  private final BooleanProperty playAsCharacterVisible;
+
   private final ListProperty<Character> characters;
 
   private final ClientCharacterSheet client;
-  private final ICharacterSheetModel characterSheetModel;
   private final IntegerProperty indexProperty;
 
-  public CharacterViewModel(ICharacterSheetModel characterSheetModel,
-      ClientCharacterSheet client)
+  public CharacterViewModel(ClientCharacterSheet client)
   {
     characterName = new SimpleStringProperty();
     strength = new SimpleStringProperty();
@@ -75,105 +78,18 @@ public class CharacterViewModel
     isPlayAsCharacterDisabled = new SimpleBooleanProperty(false);
     playAsCharacterText = new SimpleStringProperty("Play as character");
 
+    removeCharacterButtonDisabled = new SimpleBooleanProperty();
+    characterInfoVisible = new SimpleBooleanProperty(false);
+    playAsCharacterVisible = new SimpleBooleanProperty(false);
+
     saveStatusText = new SimpleStringProperty();
     saveStatusColor = new SimpleObjectProperty<>(Color.BLACK);
 
     indexProperty = new SimpleIntegerProperty();
 
-    this.characterSheetModel = characterSheetModel;
     this.client = client;
 
     characters = new SimpleListProperty<>(FXCollections.observableArrayList(client.getCharacters()));
-  }
-
-  public StringProperty characterNameProperty() {
-    return characterName;
-  }
-
-  public StringProperty characterClassProperty() {
-    return characterClass;
-  }
-
-  public StringProperty levelProperty() {
-    return level;
-  }
-
-  public StringProperty strengthProperty() {
-    return strength;
-  }
-
-  public StringProperty strengthModProperty()
-  {
-    return strengthMod;
-  }
-
-  public StringProperty dexterityProperty() {
-    return dexterity;
-  }
-
-  public StringProperty dexterityModProperty()
-  {
-    return dexterityMod;
-  }
-
-  public StringProperty constitutionProperty() {
-    return constitution;
-  }
-
-  public StringProperty constitutionModProperty()
-  {
-    return constitutionMod;
-  }
-
-  public StringProperty intelligenceProperty() {
-    return intelligence;
-  }
-
-  public StringProperty intelligenceModProperty()
-  {
-    return intelligenceMod;
-  }
-
-  public StringProperty wisdomProperty() {
-    return wisdom;
-  }
-
-  public StringProperty wisdomModProperty()
-  {
-    return wisdomMod;
-  }
-
-  public StringProperty charismaProperty() {
-    return charisma;
-  }
-
-  public StringProperty charismaModProperty()
-  {
-    return charismaMod;
-  }
-
-  public StringProperty maxHpProperty() {
-    return maxHp;
-  }
-
-  public BooleanProperty isPlayAsCharacterDisabledProperty() {
-    return isPlayAsCharacterDisabled;
-  }
-
-  public StringProperty playAsCharacterTextProperty() {
-    return playAsCharacterText;
-  }
-
-  public StringProperty saveStatusTextProperty() {
-    return saveStatusText;
-  }
-
-  public ObjectProperty<Paint> saveStatusColorProperty() {
-    return saveStatusColor;
-  }
-
-  public ListProperty<Character> charactersProperty() {
-    return characters;
   }
 
   private String setModStat(String stat)
@@ -200,7 +116,8 @@ public class CharacterViewModel
       saveStatusText.setValue("All values must be set before saving");
       return;
     }
-    characterSheetModel.makeCharacter(
+
+    client.makeCharacter(
         characterName.getValue(),
         Integer.parseInt(strength.getValue()),
         Integer.parseInt(dexterity.getValue()),
@@ -211,6 +128,7 @@ public class CharacterViewModel
         Integer.parseInt(level.getValue()),
         characterClass.getValue(),
         Integer.parseInt(maxHp.getValue()));
+
     characters.clear();
     ArrayList<Character> newCharacters = client.getCharacters();
     characters.addAll(newCharacters);
@@ -220,6 +138,7 @@ public class CharacterViewModel
   }
 
   public void updateCharacterInfo(Character character) {
+    characterInfoVisible.set(true);
     saveStatusText.setValue("");
 
     characterName.set(character.getName());
@@ -247,31 +166,46 @@ public class CharacterViewModel
     maxHp.set("");
 
     indexProperty.set(-1);
+    characterInfoVisible.set(true);
   }
 
   public void playAsCharacter(Character character) {
-    client.setCurrentCharacter(character);
+    if (character != null) {
+      client.setCurrentCharacter(character);
+    }
   }
 
   public void updatePlayAsCharacterButton(Character character) {
-    if (character.equals(client.getUserID().getCurrentCharacter())) {
-      isPlayAsCharacterDisabled.set(true);
-      playAsCharacterText.set("Playing as character");
-    }
-    else {
-      isPlayAsCharacterDisabled.set(false);
-      playAsCharacterText.set("Play as character");
+    if (character != null) {
+      if (character.equals(client.getUserID().getCurrentCharacter())) {
+        isPlayAsCharacterDisabled.set(true);
+        playAsCharacterText.set("Playing as character");
+      }
+      else {
+        isPlayAsCharacterDisabled.set(false);
+        playAsCharacterText.set("Play as character");
+      }
+      playAsCharacterVisible.set(true);
     }
   }
 
   public void removeCharacter(Character character){
-    client.deleteCharacter(character);
-    characters.clear();
-    characters.addAll(client.getCharacters());
+    if (character != null) {
+      client.deleteCharacter(character);
+      characters.clear();
+      characters.addAll(client.getCharacters());
+    }
   }
 
-  public IntegerProperty indexProperty() {
-    return indexProperty;
+  public void bindBidirectionalIndexProperty(
+      MultipleSelectionModel<Character> selectionModel){
+    selectionModel.selectedIndexProperty().addListener((observableValue, oldIndex, newIndex) -> {
+      if (newIndex.intValue() != indexProperty.get()) indexProperty.setValue(newIndex.intValue());
+    });
+
+    indexProperty.addListener((observableValue, oldNumber, newNumber) -> {
+      if (selectionModel.getSelectedIndex() != newNumber.intValue()) selectionModel.clearAndSelect(newNumber.intValue());
+    });
   }
 
   public void onTableSelectionChanged(ObservableValue<? extends Character> obs, Character oldValue,Character newValue){
@@ -374,5 +308,17 @@ public class CharacterViewModel
 
   public ListProperty<Character> charactersProperty() {
     return characters;
+  }
+
+  public BooleanProperty removeCharacterButtonDisabledProperty() {
+    return removeCharacterButtonDisabled;
+  }
+
+  public BooleanProperty characterInfoVisibleProperty() {
+    return characterInfoVisible;
+  }
+
+  public BooleanProperty playAsCharacterVisibleProperty() {
+    return playAsCharacterVisible;
   }
 }
